requirements.txt
pytorch=0.4.1